src/rest/requirements.txt
amqp==5.0.2
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
billiard==3.6.3.0
bleach==3.2.1
celery==5.0.5
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djongo==1.3.3
entrypoints==0.3
idna==2.10
inflect==4.1.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.1
jupyterlab-server==1.2.0
kiwisolver==1.3.1
kombu==5.0.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.4
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
numpy==1.19.2
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pymongo==3.11.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
redis==3.5.3
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.2.4
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.4
urllib3==1.25.10
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
gunicorn==20.0.4